Double commit of PR27973 and PR27875 (#27981)
* SPP core: changes to integrate with HPE (#27875)

* SPP core: changes to integrate with HPE

* SPP win: fix the dll API name, use oem_

* 1.14 e spp error handling (#27973)

* SPP core: add more error code for incompatible drivers, mode etc.
SPP linux, win, ESX: fix the error codes.

* SPP penutil: print success or fail

* SPP linux: print errors if update requested over adminQ

* SPP ESX: use SPP error codes

* SPP esx: new error code in case channle failed

* SPP win: fix error codes as per Val's suggestion

* SPP: packaging script improvement

* SPP esx: fix platform_init error

+++ platform/tools/package-penutil.sh
@@ -41,9 +41,14 @@ mkdir -p "$GEN_LINUX_DIR"
 #
 #LD_LIBRARY_PATH=$LINUX_PACKAGE $LINUX_PACKAGE/penutil -c $VERSION -p $FW_GEN_DIR
 
-cp $TOP/nic/dsc_fw_${VERSION}.tar $FW_GEN_DIR/
+if [ -f "$TOP/nic/dsc_fw_${VERSION}.tar" ]; then
+	cp $TOP/nic/dsc_fw_${VERSION}.tar $FW_GEN_DIR/
+fi
 
-cp $PLAT_GEN_DIR/penutil-windows.zip $GEN_DIR/
+if [ -f "$PLAT_GEN_DIR/penutil-windows.zip" ]; then
+	cp $PLAT_GEN_DIR/penutil-windows.zip $GEN_DIR/
+fi
+
 mv $ESXI67_PACKAGE/*penutil*.* $GEN_DIR/
 mv $ESXI70_PACKAGE/*penutil*.* $GEN_DIR/
